+++ api/controllers/messages/get-web.js
@@ -15,7 +15,7 @@ module.exports = {
 
         try {
             // Get the client IP address
-            var from_IP = sails.helpers.getIP(this.req);
+            var from_IP = await sails.helpers.getIp(this.req);
             var host = sh.unique(from_IP + sails.config.custom.hostSecret);
 
             if (this.req.isSocket)