+++ api/controllers/songs/get.js
@@ -57,7 +57,7 @@ module.exports = {
         try {
 
             // Get the client IP address
-            var from_IP = sails.helpers.getIP(this.req);
+            var from_IP = await sails.helpers.getIp(this.req);
 
             var subcatIDs = [];
             var cats = {};