+++ api/helpers/get-ip.js
@@ -1,7 +1,7 @@
 module.exports = {
 
 
-  friendlyName: 'sails.helpers.getIP()',
+  friendlyName: 'sails.helpers.getIp()',
 
 
   description: 'Determine what the IP address is of a request',
@@ -16,13 +16,8 @@ module.exports = {
   },
 
 
-  exits: {
-
-  },
-
-
   fn: async function (inputs, exits) {
-      
+
       return exits.success(inputs.req.isSocket ? (typeof inputs.req.socket.handshake.headers['x-forwarded-for'] !== 'undefined' ? inputs.req.socket.handshake.headers['x-forwarded-for'] : inputs.req.socket.conn.remoteAddress) : inputs.req.ip);
 
   }