+++ CHANGELOG.md
@@ -6,13 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 ## Added
- - "device" parameter to recipients. This will contain the OneSignal UUID if the recipient is using the WWSU mobile app. Otherwise, it will be null.
+ - "device" property returned for recipient objects in recipients/get. This will contain the OneSignal UUID if the recipient is using the WWSU mobile app. Otherwise, it will be null.
+ - "device" optional parameter for requests/place. If provided, the onesignal ID will receive a push notification when the requested track is played.
+ - "device=" query-string parameter for /listen and all /listen/* web pages. Specify OneSignal ID if loading these pages from the WWSU app. That way, subscribing / push notifications will be enabled for that device.
+ - "device" optional parameter for recipients/add-web . Provide the OneSignal ID if the recipient is from the WWSU mobile app.
+ - "device" optional parameter in sails.helpers.recipients.add. Provide the OneSignal ID if the recipient is from the WWSU mobile app.
 
 ### Changed
  - Many of the methods used in api controllers and in models have been migrated to sails helpers.
  - Removed a lot of the logging that could potentially output sensitive data to the console.
  - "calendar" model now uses nodebase datastore instead of RAM. This is so we can persist-store event cancellations (which would not be done through Google Calendar).
  - "active" parameter of calendar events will now be false either if the event passed the end time, or if the event was marked as canceled.
+ - requests/place response object changed. HTML property removed; instead, message property will be used, which will NOT contain div/bootstrap data.
+ - sails.helpers.requests.checkRequestable now returns message property instead of HTML property; message property does not contain div/bootstrap data.
 
 ## [5.1.0] - 2019-03-15
 ### Deprecated